# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "EasyPeasy API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/grawls": {
      "get": {
        "tags": [
          "GrawlApi"
        ],
        "operationId": "GetGrawls",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Grawl"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GrawlApi"
        ],
        "operationId": "CreateGrawl",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Grawl"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grawl"
            }
          }
        }
      },
      "put": {
        "tags": [
          "GrawlApi"
        ],
        "operationId": "EditGrawl",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Grawl"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grawl"
            }
          }
        }
      }
    },
    "/api/grawls/{id}": {
      "get": {
        "tags": [
          "GrawlApi"
        ],
        "operationId": "GetGrawl",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grawl"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GrawlApi"
        ],
        "operationId": "DeleteGrawl",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/grawls/{name}": {
      "get": {
        "tags": [
          "GrawlApi"
        ],
        "operationId": "GetGrawlByName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Grawl"
            }
          }
        }
      }
    },
...
    "/api/grawltasks": {
      "get": {
        "tags": [
          "GrawlTaskApi"
        ],
        "operationId": "GetGrawlTasks",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GrawlTask"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GrawlTaskApi"
        ],
        "operationId": "CreateGrawlTask",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GrawlTask"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTask"
            }
          }
        }
      },
      "put": {
        "tags": [
          "GrawlTaskApi"
        ],
        "operationId": "EditGrawlTask",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GrawlTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTask"
            }
          }
        }
      }
    },
    "/api/grawltasks/{id}": {
      "get": {
        "tags": [
          "GrawlTaskApi"
        ],
        "operationId": "GetGrawlTask",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTask"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GrawlTaskApi"
        ],
        "operationId": "DeleteGrawlTask",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/taskings": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetAllGrawlTaskings",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GrawlTasking"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "CreateGrawlTasking",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GrawlTasking"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTasking"
            }
          }
        }
      },
      "put": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "EditGrawlTasking",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GrawlTasking"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTasking"
            }
          }
        }
      }
    },
    "/api/grawls/{id}/taskings": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetGrawlTaskings",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GrawlTasking"
              }
            }
          }
        }
      }
    },
    "/api/grawls/{id}/taskings/search": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetSearchGrawlTaskings",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GrawlTasking"
              }
            }
          }
        }
      }
    },
    "/api/grawls/{id}/taskings/uninitialized": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetUninitializedGrawlTaskings",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GrawlTasking"
              }
            }
          }
        }
      }
    },
    "/api/grawls/{id}/taskings/search/uninitialized": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetSearchUninitializedGrawlTaskings",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GrawlTasking"
              }
            }
          }
        }
      }
    },
    "/api/taskings/{tid}": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetGrawlTasking",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTasking"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "DeleteGrawlTasking",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/grawls/taskings/{taskingname}": {
      "get": {
        "tags": [
          "GrawlTaskingApi"
        ],
        "operationId": "GetGrawlTaskingByName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "taskingname",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrawlTasking"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EmbeddedResource": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "location": {
          "type": "string"
        }
      }
    },
...
    "ImplantLanguage": {
      "enum": [
        "CSharp"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantLanguage",
        "modelAsString": false
      }
    },
    "CommunicationType": {
      "enum": [
        "HTTP",
        "SMB",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "CommunicationType",
        "modelAsString": false
      }
    },
...
    "DotNetVersion": {
      "enum": [
        "Net35",
        "Net40",
        "NetCore31"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "DotNetVersion",
        "modelAsString": false
      }
    },
    "ImplantTemplate": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "commType": {
          "$ref": "#/definitions/CommunicationType"
        },
        "implantDirection": {
          "$ref": "#/definitions/ImplantDirection"
        },
        "compatibleListenerTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListenerType"
          },
          "readOnly": true
        },
        "compatibleDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "stagerCode": {
          "type": "string"
        },
        "executorCode": {
          "type": "string"
        }
      }
    },
...
    "Listener": {
      "required": [
        "bindAddress",
        "bindPort",
        "connectAddresses",
        "connectPort",
        "description",
        "anotherid",
        "listenerTypeId",
        "name",
        "profileId",
        "status"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "anotherid": {
          "maxLength": 100,
          "minLength": 0,
          "pattern": "^[a-zA-Z0-9]*$",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "bindAddress": {
          "pattern": "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$",
          "type": "string"
        },
        "bindPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "connectAddresses": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "connectPort": {
          "format": "int32",
          "maximum": 65535,
          "minimum": 1,
          "type": "integer"
        },
        "profileId": {
          "format": "int32",
          "type": "integer"
        },
        "profile": {
          "$ref": "#/definitions/Profile"
        },
        "listenerTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "listenerType": {
          "$ref": "#/definitions/ListenerType"
        },
        "status": {
          "$ref": "#/definitions/ListenerStatus"
        },
        "covenantUrl": {
          "type": "string"
        },
        "covenantToken": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "RuntimeIdentifier": {
      "enum": [
        "win_x64",
        "win_x86",
        "win_arm",
        "win_arm64",
        "win7_x64",
        "win7_x86",
        "win81_x64",
        "win81_x86",
        "win81_arm",
        "win10_x64",
        "win10_x86",
        "win10_arm",
        "win10_arm64",
        "linux_x64",
        "linux_musl_x64",
        "linux_arm",
        "linux_arm64",
        "rhel_x64",
        "rhel_6_x64",
        "tizen",
        "tizen_4_0_0",
        "tizen_5_0_0",
        "osx_x64",
        "osx_10_10_x64",
        "osx_10_11_x64",
        "osx_10_12_x64",
        "osx_10_13_x64",
        "osx_10_14_x64",
        "osx_10_15_x64"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "RuntimeIdentifier",
        "modelAsString": false
      }
    },
    "GrawlStatus": {
      "enum": [
        "Uninitialized",
        "Stage0",
        "Stage1",
        "Stage2",
        "Active",
        "Lost",
        "Exited",
        "Disconnected",
        "Hidden"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "GrawlStatus",
        "modelAsString": false
      }
    },
    "IntegrityLevel": {
      "enum": [
        "Untrusted",
        "Low",
        "Medium",
        "High",
        "System"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "IntegrityLevel",
        "modelAsString": false
      }
    },
    "GrawlTaskAuthor": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "handle": {
          "type": "string"
        },
        "link": {
          "type": "string"
        },
        "grawlTasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GrawlTask"
          }
        }
      }
    },
    "GrawlTaskingType": {
      "enum": [
        "Assembly",
        "SetDelay",
        "SetJItter",
        "SetConneCTAttEmpts",
        "SetKillDate",
        "Exit",
        "Connect",
        "Disconnect",
        "Tasks",
        "TaskKill"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "GrawlTaskingType",
        "modelAsString": false
      }
    },
    "ReferenceAssembly": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "dotNetVersion": {
          "$ref": "#/definitions/DotNetVersion"
        }
      }
    },
    "ReferenceSourceLibrary": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "compatibleDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "referenceAssemblies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceAssembly"
          },
          "readOnly": true
        },
        "embeddedResources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmbeddedResource"
          },
          "readOnly": true
        }
      }
    },
...
    "GrawlTask": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "authorId": {
          "format": "int32",
          "type": "integer"
        },
        "author": {
          "$ref": "#/definitions/GrawlTaskAuthor"
        },
        "name": {
          "type": "string"
        },
        "aliases": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": {
          "type": "string"
        },
        "help": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "compatibleDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "code": {
          "type": "string"
        },
        "compiled": {
          "type": "boolean"
        },
        "taskingType": {
          "$ref": "#/definitions/GrawlTaskingType"
        },
        "referenceSourceLibraries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceSourceLibrary"
          },
          "readOnly": true
        },
        "referenceAssemblies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReferenceAssembly"
          },
          "readOnly": true
        },
        "embeddedResources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EmbeddedResource"
          },
          "readOnly": true
        },
        "unsafeCompile": {
          "type": "boolean"
        },
        "tokenTask": {
          "type": "boolean"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GrawlTaskOption"
          }
        }
      }
    },
    "GrawlTaskingStatus": {
      "enum": [
        "Uninitialized",
        "Tasked",
        "Progressed",
        "Completed",
        "Aborted"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "GrawlTaskingStatus",
        "modelAsString": false
      }
    },
    "GrawlTasking": {
      "required": [
        "grawlId",
        "grawlTaskId",
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "grawlId": {
          "format": "int32",
          "type": "integer"
        },
        "grawl": {
          "$ref": "#/definitions/Grawl"
        },
        "grawlTaskId": {
          "format": "int32",
          "type": "integer"
        },
        "grawlTask": {
          "$ref": "#/definitions/GrawlTask"
        },
        "type": {
          "$ref": "#/definitions/GrawlTaskingType"
        },
        "parameters": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "$ref": "#/definitions/GrawlTaskingStatus"
        },
        "taskingTime": {
          "format": "date-time",
          "type": "string"
        },
        "completionTime": {
          "format": "date-time",
          "type": "string"
        },
        "grawlCommandId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "GrawlCommand": {
      "required": [
        "command",
        "commandOutputId",
        "commandTime",
        "userId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "command": {
          "type": "string"
        },
        "commandTime": {
          "format": "date-time",
          "type": "string"
        },
        "commandOutputId": {
          "format": "int32",
          "type": "integer"
        },
        "commandOutput": {
          "$ref": "#/definitions/CommandOutput"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/EasyPeasyUser"
        },
        "grawlTaskingId": {
          "format": "int32",
          "type": "integer"
        },
        "grawlTasking": {
          "$ref": "#/definitions/GrawlTasking"
        },
        "grawlId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Grawl": {
      "required": [
        "connectAttempts",
        "delay",
        "dotNetVersion",
        "implantTemplateId",
        "integrity",
        "jitterPercent",
        "killDate",
        "name",
        "originalServerGuid",
        "runtimeIdentifier",
        "smbPipeName",
        "status",
        "useCertPinning",
        "validateCert"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "originalServerGuid": {
          "type": "string"
        },
        "anotherid": {
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "implantTemplateId": {
          "format": "int32",
          "type": "integer"
        },
        "implantTemplate": {
          "$ref": "#/definitions/ImplantTemplate"
        },
        "validateCert": {
          "type": "boolean"
        },
        "useCertPinning": {
          "type": "boolean"
        },
        "smbPipeName": {
          "type": "string"
        },
        "listenerId": {
          "format": "int32",
          "type": "integer"
        },
        "listener": {
          "$ref": "#/definitions/Listener"
        },
        "note": {
          "type": "string"
        },
        "delay": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 0,
          "type": "integer"
        },
        "jitterPercent": {
          "format": "int32",
          "maximum": 100,
          "minimum": 0,
          "type": "integer"
        },
        "connectAttempts": {
          "format": "int32",
          "maximum": 2147483647,
          "minimum": 0,
          "type": "integer"
        },
        "killDate": {
          "format": "date-time",
          "type": "string"
        },
        "dotNetVersion": {
          "$ref": "#/definitions/DotNetVersion"
        },
        "runtimeIdentifier": {
          "$ref": "#/definitions/RuntimeIdentifier"
        },
        "status": {
          "$ref": "#/definitions/GrawlStatus"
        },
        "integrity": {
          "$ref": "#/definitions/IntegrityLevel"
        },
        "process": {
          "type": "string"
        },
        "userDomainName": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "ipAddress": {
          "type": "string"
        },
        "hostname": {
          "type": "string"
        },
        "operatingSystem": {
          "type": "string"
        },
        "grawlSharedSecretPassword": {
          "type": "string"
        },
        "grawlRSAPublicKey": {
          "type": "string"
        },
        "grawlNegotiatedSessKEy": {
          "type": "string"
        },
        "grawlChallenge": {
          "type": "string"
        },
        "activationTime": {
          "format": "date-time",
          "type": "string"
        },
        "lastCheckIn": {
          "format": "date-time",
          "type": "string"
        },
        "powerShellImport": {
          "type": "string"
        },
        "grawlCommands": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GrawlCommand"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}